pyproject.toml
[tool.poetry]
name = "ct-iot-authentication-service"
version = "0.0.1"
description = "ct-iot-authentication-service"
authors = ["Darryl Morton"]
license = "GNU"
readme = "README.md"
packages = [{include = "authentication_service", from="src"}]
repository = "https://github.com/darrylmorton/ct-iot-authentication-service.git"
keywords = ["ct", "iot", "authentication", "service"]

[tool.poetry.dependencies]
python = "^3.11.9"
fastapi = "^0.111.0"
structlog = "^24.2.0"
python-dotenv = "^1.0.1"
toml = "^0.10.2"
sentry-sdk = {version = "^2.7.1", extras = ["fastapi"]}
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
async-sqlalchemy = "^1.0.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
psycopg = "^3.2.1"
sqlalchemy = "^2.0.31"
bcrypt = "^4.1.3"
greenlet = "^3.0.3"
requests = "^2.32.3"


[tool.poetry.group.test.dependencies]
#httpx = "^0.25.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.7"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
preview = true
include = ["*.py", "*.pyi", "pyproject.toml", "*.ipynb"]
line-length = 88  # defaults to 88 like black
target-version = "py312"  # the python version to target, useful when considering code upgrades, defaults to "py310"
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F", "E"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
#magic-trailing-comma = "respect"

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["src/"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "--cov --cov-config=./tests/pytest-cov.ini --cov-report term-missing --cov-report=html"
